fix(pg): repair 3 main-red tests (force-transition / review-tuning) broken by INT4→BIGINT migration (#947)
* fix(pg): repair 2 main-red tests broken by INT4→BIGINT migration

Scope: L10081 (postgres_force_transition_to_ready_cleans_up_live_state)
and L10767 (transition_to_done_records_true_negative_in_postgres_review_tuning).
Both were silently failing because `card_review_state.review_round` became
BIGINT in `0008_int4_to_bigint_audit.sql` but sqlx call sites still decoded
as `i32`. The second site further swallowed the resulting decode error
inside `block_on_pg_result(..).unwrap_or(false)`, which aborted the entire
PG true_negative recording path — hence the RowNotFound panic downstream.

Changes:
- `routes_tests.rs:10067,10083` — decode `kanban_cards.review_round` and
  `card_review_state.review_round` as `i64` (post-0008).
- `kanban.rs::record_true_negative_if_pass_with_backends` — decode as
  `Option<i64>`, then downcast to `i32` at the INSERT site because
  `review_tuning_outcomes.review_round` was NOT included in the 0008
  audit and remains INTEGER.

L9621 (`force_transition_to_ready_cancels_live_dispatches_and_skips_auto_queue_entries`)
is NOT included here. Making it pass requires extending the
`set_dispatch_status_*` side-effect to clear `dispatch_id` when
terminalizing linked auto_queue_entries, which ripples into 20+ cancel
route tests — too large for this PR. Tracked separately.

Same fault family as fix(pg) 8de3842f / 4c874c5c / 3399edc5.

* chore: regenerate inventory docs to clear Script checks drift

Line-number drift accumulated on main since last inventory regeneration.
This keeps PR #947 CI green by aligning docs/generated with the current
source layout. No source code changes.

---------

+++ src/server/routes/routes_tests.rs
@@ -10064,7 +10064,7 @@ async fn postgres_force_transition_to_ready_cleans_up_live_state() {
         String,
         Option<String>,
         Option<String>,
-        i32,
+        i64,
         Option<String>,
         Option<String>,
         Option<String>,
@@ -10080,7 +10080,7 @@ async fn postgres_force_transition_to_ready_cleans_up_live_state() {
     .await
     .unwrap();
     let (review_state_round, review_state_status, review_state_pending_dispatch): (
-        i32,
+        i64,
         String,
         Option<String>,
     ) = sqlx::query_as(
